refactor!: Overhaul status update mutations
Breaking Change: The manual handling of the status update streaks has
been removed in anticipation of moving the streak logic to the business
layer instead of the client.

In its place, the API will only expose a singular mutation to mark a
particular day's status update status for the member.